Refactor main for cleanliness
